backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
azure-ai-inference==1.0.0b7
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
azure-storage-blob==12.23.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.12.0
googletrans==4.0.0rc1
gTTS==2.5.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
openai==1.60.0
pandas==2.0.3
pillow==10.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.0
PyMuPDF==1.24.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
